+++ src/Common/EngineCore/Systems/WorldManagement/Components/Indexers/NonBlockWorldSegmentIndexer.cs
@@ -69,7 +69,7 @@ public class NonBlockWorldSegmentIndexer : BaseComponentIndexer<Vector3>
         return new HashSet<ulong>(index[segmentIndex]);
     }
 
-    protected override void ComponentAddedCallback(ulong entityId, Vector3 componentValue)
+    protected override void ComponentAddedCallback(ulong entityId, Vector3 componentValue, bool isLoad)
     {
         AddEntity(entityId, componentValue);
     }
@@ -85,12 +85,7 @@ public class NonBlockWorldSegmentIndexer : BaseComponentIndexer<Vector3>
         }
     }
 
-    protected override void ComponentRemovedCallback(ulong entityId)
-    {
-        RemoveEntity(entityId);
-    }
-
-    protected override void ComponentUnloadedCallback(ulong entityId)
+    protected override void ComponentRemovedCallback(ulong entityId, bool isUnload)
     {
         RemoveEntity(entityId);
     }